Cpalms_indexing/requirements.txt
# Azure Functions
azure-functions>=1.18.0

# Azure SDK packages
azure-search-documents==11.4.0
azure-storage-blob==12.19.0
azure-core==1.29.5
azure-identity==1.15.0

# OpenAI - Updated to latest stable version
openai>=1.30.0

# Database
sqlalchemy==2.0.25
pyodbc==5.0.1

# Environment variables
python-dotenv==1.0.0

# HTTP requests
requests==2.31.0
aiohttp>=3.8.0

# Explicitly pin compatible httpx version
httpx>=0.24.0,<1.0.0
tiktoken==0.7.0
